remove index, refactor tests, add docs
