+++ crates/cell/Cargo.toml
@@ -35,6 +35,8 @@ mockall = "0.6.0"
 tokio = { version = "0.2.11", features = ["macros", "rt-threaded", "rt-util", "sync"] }
 owning_ref = "0.4.1"
 anyhow = "1.0.26"
+tracing = "0.1.13"
 
 [dev-dependencies]
 anyhow = "1.0.26"
+test_utils = { path = "../test_utils" }
