+++ crates/state/Cargo.toml
@@ -34,6 +34,7 @@ bincode = "1.2.1"
 
 [dev-dependencies]
 anyhow = "1.0.26"
+test_utils = { path = "../test_utils" }
 
 # TODO: this brings in too many assumptions from the previous version,
 # but there is a lot of useful stuff here, so will have to refactor all of this later
