+++ crates/state/src/lib.rs
@@ -10,4 +10,4 @@ pub mod transaction;
 
 // NB: would be nice to put this under cfg(test), but then it's not visible from other crates,
 // since cfg(test) only applies to the crate in which you run tests
-pub mod test_utils;
+//pub mod test_utils;
